Move Facebook/passport setup from main.js to passportSetup module
